Refactor: 기존 모달 컴포넌트 수정
